requirements.txt
# A2A Framework
a2a-sdk>=0.3.5
google-adk>=0.1.0
protobuf>=5.0,<6

# AgentBeats Controller (for local development and testing)
earthshaker>=0.1.0

# Google GenAI
google-genai>=1.0.0

# LLM Providers
groq>=1.0.0
litellm>=1.75.5

# Server
uvicorn[standard]>=0.30.0
starlette>=0.40.0

# Data/Config
pydantic>=2.0.0
python-dotenv>=1.0.0
toml>=0.10.0

# HTTP
httpx>=0.27.0

# Logging
python-json-logger>=2.0.0
